+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Change Log for SD.Next
 
-## Update for 2025-07-10
+## Update for 2025-07-11
 
-### Highlights for 2025-07-10
+### Highlights for 2025-07-11
 
 In this release we finally break with legacy with the removal of the original [A1111](https://github.com/AUTOMATIC1111/stable-diffusion-webui/) codebase which has not been maintained for a while now  
 This plus major cleanup of codebase and external dependencies resulted in ~53k LoC (*lines-of-code*) reduction and spread over [~680 files](https://github.com/vladmandic/sdnext/pull/4017)!  
@@ -23,7 +23,7 @@ Although upgrades and existing installations are tested and should work fine!
 
 [ReadMe](https://github.com/vladmandic/automatic/blob/master/README.md) | [ChangeLog](https://github.com/vladmandic/automatic/blob/master/CHANGELOG.md) | [Docs](https://vladmandic.github.io/sdnext-docs/) | [WiKi](https://github.com/vladmandic/automatic/wiki) | [Discord](https://discord.com/invite/sd-next-federal-batch-inspectors-1101998836328697867)
 
-### Details for 2025-07-10
+### Details for 2025-07-11
 
 - **License**  
   - SD.Next [license](https://github.com/vladmandic/sdnext/blob/dev/LICENSE.txt) switched from **aGPL-v3.0** to **Apache-v2.0**  
@@ -88,6 +88,8 @@ Although upgrades and existing installations are tested and should work fine!
   - improve model type autodetection  
   - improve model auth check for hf repos  
   - improve Chroma prompt padding as per recommendations  
+  - lock directml torch to `torch-directml==0.2.4.dev240913`  
+  - lock directml transformers to `transformers==4.52.4`  
 - **Refactoring**  
   *note*: none of the removals result in loss-of-functionality since all those features are already re-implemented  
   goal here is to remove legacy code, code duplication and reduce code complexity